% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_composition.R
\name{adult_cows_body_composition}
\alias{adult_cows_body_composition}
\title{Body composition of adult cows.}
\usage{
adult_cows_body_composition(body_weight, BCS)
}
\arguments{
\item{body_weight}{Body weight (kg).}

\item{BCS}{Body condition score (1 to 5).}
}
\value{
Proportion of protein, fat, water, and ash in the empty body weight of adult cows.
}
\description{
Body composition of adult cows.
}
\examples{
adult_cows_body_composition(body_weight = 680, BCS = 3)
}
\seealso{
The equation used here is also available in \href{https://nap.nationalacademies.org/catalog/9825/nutrient-requirements-of-dairy-cattle-seventh-revised-edition-2001}{NRC (2001)}.
}
