% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_carbon_dioxide.R
\name{animal_co2_emissions}
\alias{animal_co2_emissions}
\title{Estimates de daily carbon dioxide respired emissions from the animals.}
\usage{
animal_co2_emissions(dry_matter_intake, body_weight)
}
\arguments{
\item{dry_matter_intake}{Dry matter intake (kg).}

\item{body_weight}{Body weight (kg).}
}
\value{
Daily carbon dioxide emissions (kg).
}
\description{
Estimates de daily carbon dioxide respired emissions from the animals.
}
\examples{
animal_co2_emissions(dry_matter_intake = 25, body_weight = 680)

}
\seealso{
The equation used here is also available in \href{https://www.osti.gov/etdeweb/biblio/6065141}{Kirchgessner et al. (1991)}.
}
