% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_nitrous_oxide.R
\name{animal_n2o_emissions}
\alias{animal_n2o_emissions}
\title{Estimates the nitrous oxide emissions from the animals.}
\usage{
animal_n2o_emissions(dry_matter_intake, crude_protein)
}
\arguments{
\item{dry_matter_intake}{Dry matter intake (kg).}

\item{crude_protein}{Diet crude protein (\%).}
}
\value{
Daily nitrous oxide emissions (g).
}
\description{
Estimates the nitrous oxide emissions from the animals.
}
\examples{
animal_n2o_emissions(dry_matter_intake = 25, crude_protein = 16)
}
\seealso{
The equation used here is also available in \href{https://www.ars.usda.gov/northeast-area/up-pa/pswmru/docs/dairy-gas-emissions-model/}{Rotz et al. (2016)}.
}
