% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_inputs.R
\name{animal_ui_prms}
\alias{animal_ui_prms}
\title{Animal UI parameters.}
\usage{
animal_ui_prms(prm)
}
\arguments{
\item{prm}{Parameter to be concatenated in the module.}
}
\value{
Animal UI components.
}
\description{
Animal UI parameters.
}
