% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodigester_methane.R
\name{biodigester_ch4_yield}
\alias{biodigester_ch4_yield}
\title{Estimates the methane yield from the biodigester.}
\usage{
biodigester_ch4_yield(volatile_solids, biodigester_efficiency)
}
\arguments{
\item{volatile_solids}{Total volatile solids in the afluent (kg).}

\item{biodigester_efficiency}{Biodigester efficiency (values vary from 20\% to 45\%)}
}
\value{
Total methane yield (kg).
}
\description{
Estimates the methane yield from the biodigester.
}
\examples{
biodigester_ch4_yield(volatile_solids = 2088, biodigester_efficiency = 30)

}
\seealso{
The equation used here is also available in \href{https://www.ars.usda.gov/northeast-area/up-pa/pswmru/docs/integrated-farm-system-model/#Reference}{Rotz et al. (2022)}.
}
