% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lactating_cow_body_weight.R
\name{body_weight_loss_lactation}
\alias{body_weight_loss_lactation}
\title{Estimates the body weight losses daily over the lactation.}
\usage{
body_weight_loss_lactation(days_milk, parity)
}
\arguments{
\item{days_milk}{Days in milk.}

\item{parity}{Parity: primiparous or multiparous.}
}
\value{
Body weight loss at given day in milk (kg).
}
\description{
Estimates the body weight losses daily over the lactation.
}
\examples{
body_weight_loss_lactation(days_milk = 30, parity = "multiparous")
}
\seealso{
The equations used here are also available in \href{https://pubmed.ncbi.nlm.nih.gov/23415521/}{Galvão et al. (2013)}.
}
