% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/built_plots.R
\name{built_gauge_plotly}
\alias{built_gauge_plotly}
\title{Built gauge using plotly.}
\usage{
built_gauge_plotly(carbon_equivalent)
}
\arguments{
\item{carbon_equivalent}{Carbon equivalent.}
}
\value{
Gauge.
}
\description{
Built gauge using plotly.
}
