% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/built_plots.R
\name{built_variable_over_lactation_plot}
\alias{built_variable_over_lactation_plot}
\title{Built ggplot2 plot for variables over the lactation.}
\usage{
built_variable_over_lactation_plot(
  data_frame,
  x_variable,
  y_variable,
  col_variable,
  y_title
)
}
\arguments{
\item{data_frame}{Data frame.}

\item{x_variable}{X axis variable.}

\item{y_variable}{Y axis variable.}

\item{col_variable}{Color variable.}

\item{y_title}{Y axis title.}
}
\value{
Ggplot2 plot.
}
\description{
Built ggplot2 plot for variables over the lactation.
}
