% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/built_plots.R
\name{built_variable_over_year_plot}
\alias{built_variable_over_year_plot}
\title{Built ggplot2 over the year.}
\usage{
built_variable_over_year_plot(
  data_frame,
  x_variable,
  y_variable,
  y_title,
  plot_title
)
}
\arguments{
\item{data_frame}{Data frame.}

\item{x_variable}{X axis variable.}

\item{y_variable}{Y xis variable.}

\item{y_title}{Y axis title.}

\item{plot_title}{Plot title.}
}
\value{
Ggplor2 plot.
}
\description{
Built ggplot2 over the year.
}
