% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/built_plots.R
\name{built_waterfall_plot}
\alias{built_waterfall_plot}
\title{Built waterfall plot using ggplot2.}
\usage{
built_waterfall_plot(data_frame, x_axis, y_axis)
}
\arguments{
\item{data_frame}{Data frame.}

\item{x_axis}{X axis title.}

\item{y_axis}{Y axis title.}
}
\value{
Waterfall plot.
}
\description{
Built waterfall plot using ggplot2.
}
