% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/built_plots.R
\name{built_waterfall_plot_plotly}
\alias{built_waterfall_plot_plotly}
\title{Built waterfall plot using plotly.}
\usage{
built_waterfall_plot_plotly(data_frame, nutrient_balance, x_axis_title)
}
\arguments{
\item{data_frame}{Data frame.}

\item{nutrient_balance}{Nutrient balance.}

\item{x_axis_title}{X axis title.}
}
\value{
Waterfall plot Plotly.
}
\description{
Built waterfall plot using plotly.
}
