% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calf_adg.R
\name{calf_adg}
\alias{calf_adg}
\title{Estimates the average calf daily gain.}
\usage{
calf_adg(milk_intake, starter_intake)
}
\arguments{
\item{milk_intake}{Calf milk intake (kg).}

\item{starter_intake}{Average starter intake (kg).}
}
\value{
Calf average daily gain in a given day (kg).
}
\description{
Estimates the average calf daily gain.
}
\examples{
# Estimate the average daily gain of a calf consuming 6 l/day of milk and 0.2 kg/day of starter
calf_adg(milk_intake = 6, starter_intake = 0.2)
}
\seealso{
The equation used here is also available in \href{https://www.simleite.com.br/arquivosAnais/arquivo118}{Silva et al. (2019)}.#TODO
}
