% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calf_adg.R
\name{calf_body_weight}
\alias{calf_body_weight}
\title{Estimates the calf body weight.}
\usage{
calf_body_weight(birth_weight, calf_adg, day_max, type)
}
\arguments{
\item{birth_weight}{Birth weight (kg).}

\item{calf_adg}{Calf average daily gain (kg).}

\item{day_max}{Day limit for weight estimation.}

\item{type}{Weight type to be estimated: average weight in the period or final weight (kg).}
}
\value{
Average calf weight or final calf weight (kg).
}
\description{
Estimates the calf body weight.
}
\examples{
calf_body_weight(birth_weight = 40, calf_adg = 0.5, day_max = 50, type = "final")
calf_body_weight(birth_weight = 40, calf_adg = 0.5, day_max = 50, type = "mean")

}
