% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_inputs.R
\name{calf_ui_prms}
\alias{calf_ui_prms}
\title{Calf UI parameters.}
\usage{
calf_ui_prms(prm)
}
\arguments{
\item{prm}{Parameter to be concatenated in the module.}
}
\value{
Calf UI components.
}
\description{
Calf UI parameters.
}
