% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_dioxide.R
\name{co2_emissions}
\alias{co2_emissions}
\title{Estimates de daily CO2 respired emissions from the animals.}
\usage{
co2_emissions(dry_matter_intake, body_weight)
}
\arguments{
\item{dry_matter_intake}{Dry matter intake (kg).}

\item{body_weight}{Body weight (kg).}
}
\value{
Daily CO2 emissions (kg).
}
\description{
Estimates de daily CO2 respired emissions from the animals.
}
\examples{
co2_emissions(dry_matter_intake = 25, body_weight = 680)
}
\seealso{
The equation used here is also available in \href{https://www.osti.gov/etdeweb/biblio/6065141}{Kirchgessner et al. (1991)}.
}
