% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fossil_fuel_combustion.R
\name{diesel_co2eq}
\alias{diesel_co2eq}
\title{Estimates the CO2eq emissions from gas diesel combustion.}
\usage{
diesel_co2eq(diesel_consumption)
}
\arguments{
\item{diesel_consumption}{Diesel consumption (L).}
}
\value{
CO2eq emissions (kg).
}
\description{
Estimates the CO2eq emissions from gas diesel combustion.
}
\examples{
diesel_co2eq(diesel_consumption = 10)
}
