% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{diet_ui_prms}
\alias{diet_ui_prms}
\title{Diet UI parameters.}
\usage{
diet_ui_prms(
  category = "diet_lac",
  cp = 16,
  ndf = 25,
  adf = 15,
  ee = 4,
  p = 0.4,
  k = 1.4
)
}
\arguments{
\item{category}{Animal categories: lactation (lac), dry (dry), and heifers(hei).}

\item{cp}{Diet crude protein (\%).}

\item{ndf}{Diet neutral detergent fiber (\%).}

\item{adf}{Diet acid detergent fiber (\%).}

\item{ee}{Diet ether extract (\%).}

\item{p}{Diet phosphorous content (\%).}

\item{k}{Diet potassium content (\%).}
}
\value{
Diet UI parameters.
}
\description{
Diet UI parameters.
}
