% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volatile_solids.R
\name{digestible_volatile_solids}
\alias{digestible_volatile_solids}
\title{Estimates the digestible fraction of the volatile solids.}
\usage{
digestible_volatile_solids(
  dry_matter_intake,
  crude_protein,
  neutral_detergent_fiber,
  acid_detergent_fiber
)
}
\arguments{
\item{dry_matter_intake}{Dry matter intake (kg).}

\item{crude_protein}{Diet crude protein (\%).}

\item{neutral_detergent_fiber}{Diet neutral detergent fiber (\%).}

\item{acid_detergent_fiber}{Diet acid detergent fiber (\%).}
}
\value{
Digestible volatile solids (kg).
}
\description{
Estimates the digestible fraction of the volatile solids.
}
\examples{
digestible_volatile_solids(dry_matter_intake = 25,
                          crude_protein = 16,
                          neutral_detergent_fiber = 30,
                          acid_detergent_fiber = 15)

}
