% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urine_excretion.R
\name{dry_and_heifer_urine_excretion}
\alias{dry_and_heifer_urine_excretion}
\title{Estimates the urine excretion for dry cows and heifers.}
\usage{
dry_and_heifer_urine_excretion(k_intake)
}
\arguments{
\item{k_intake}{Potassium intake (g/d)}
}
\value{
Urine excretion (l/d).
}
\description{
Estimates the urine excretion for dry cows and heifers.
}
\examples{
dry_and_heifer_urine_excretion(k_intake = 200)
}
