% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dry_matter_intake.R
\name{dry_cow_dry_matter_intake}
\alias{dry_cow_dry_matter_intake}
\title{Estimates the daily dry matter intake of dry cows.}
\usage{
dry_cow_dry_matter_intake(body_weight)
}
\arguments{
\item{body_weight}{Body weight (kg).}
}
\value{
Dry cow dry matter intake (kg).
}
\description{
Estimates the daily dry matter intake of dry cows.
}
\examples{
dry_cow_dry_matter_intake(680)
}
