% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enteric_methane.R
\name{dry_enteric_methane}
\alias{dry_enteric_methane}
\title{Estimates the daily enteric methane emission for dry cows.}
\usage{
dry_enteric_methane(gross_energy_intake, ether_extract)
}
\arguments{
\item{gross_energy_intake}{Gross energy intake (MJ/day).}

\item{ether_extract}{Diet ether extract concentration (\%).}
}
\value{
Dry cow enteric methane emission (g/day).
}
\description{
Estimates the daily enteric methane emission for dry cows.
}
\examples{
dry_enteric_methane(gross_energy_intake = 221, ether_extract = 3)

}
\seealso{
The equation used here is also available in \href{https://onlinelibrary.wiley.com/doi/full/10.1111/gcb.12471}{Moraes et al. (2014)}.
}
