% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manure_excretion.R
\name{dry_manure_excretion}
\alias{dry_manure_excretion}
\title{Estimates the daily total manure excretion of dry cows.}
\usage{
dry_manure_excretion(body_weight, crude_protein, neutral_detergent_fiber)
}
\arguments{
\item{body_weight}{Body Weight (kg).}

\item{crude_protein}{Diet crude protein (\%).}

\item{neutral_detergent_fiber}{Diet neutral detergent fiber (\%).}
}
\value{
Daily total manure excretion (kg).
}
\description{
Estimates the daily total manure excretion of dry cows.
}
\examples{
dry_manure_excretion(body_weight = 650, crude_protein = 12, neutral_detergent_fiber = 40)
}
