% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrogen_excretion.R
\name{dry_n_total_excretion}
\alias{dry_n_total_excretion}
\title{Estimates the daily total nitrogen excretion of dry cows.}
\usage{
dry_n_total_excretion(nitrogen_intake)
}
\arguments{
\item{nitrogen_intake}{Nitrogen intake (g/day).}
}
\value{
Daily total nitrogen excretion (g).
}
\description{
Estimates the daily total nitrogen excretion of dry cows.
}
\examples{
dry_n_total_excretion(nitrogen_intake = 230)

}
