% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fossil_fuel_combustion.R
\name{electricity_co2eq}
\alias{electricity_co2eq}
\title{Estimates the CO2eq emissions from electricity use.}
\usage{
electricity_co2eq(electricity_consumption)
}
\arguments{
\item{electricity_consumption}{Electricity consumption (kWh).}
}
\value{
CO2eq emissions (kg).
}
\description{
Estimates the CO2eq emissions from electricity use.
}
\examples{
electricity_co2eq(electricity_consumption = 10)
}
