% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{empty_day}
\alias{empty_day}
\title{Vector to determine the manure storage emptying days.}
\usage{
empty_day(empty_time)
}
\arguments{
\item{empty_time}{Season of manure storage emptying (i.e., fall, spring).}
}
\value{
Binary vector where 1 means days of emptying.
}
\description{
Vector to determine the manure storage emptying days.
}
