% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-NOP_methane_mitigation.R
\name{enteric_ch4_reduction_3nop}
\alias{enteric_ch4_reduction_3nop}
\title{3-NOP methane mitigation.}
\usage{
enteric_ch4_reduction_3nop(nop_dose, diet_ndf, diet_crude_fat)
}
\arguments{
\item{nop_dose}{3-NOP dose (mg/DM).}

\item{diet_ndf}{Diet NDF content (\%).}

\item{diet_crude_fat}{Diet ether extract content (\%).}
}
\value{
Expected CH4 mitigated after 3-NOP inclusion in the diet (\%).
}
\description{
3-NOP methane mitigation.
}
\examples{
enteric_ch4_reduction_3nop(nop_dose = 70, diet_ndf = 33, diet_crude_fat = 4.2)

}
\seealso{
The equation used here is also available in \href{https://www.journalofdairyscience.org/article/S0022-0302(22)00710-X/pdf}{Kebreab et al. (2022)}.
}
