% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enteric_methane.R
\name{enteric_methane_heifer}
\alias{enteric_methane_heifer}
\title{Estimates the daily enteric methane emission for heifers.}
\usage{
enteric_methane_heifer(
  gross_energy_intake,
  neutral_detergent_fiber,
  body_weight
)
}
\arguments{
\item{gross_energy_intake}{Gross energy intake (MJ/day).}

\item{neutral_detergent_fiber}{Diet neutral detergent fiber concentration (\%).}

\item{body_weight}{Body weight (kg).}
}
\value{
Heifer enteric methane emission (g/day).
}
\description{
Estimates the daily enteric methane emission for heifers.
}
\examples{
enteric_methane_heifer(gross_energy_intake = 110, neutral_detergent_fiber = 40, body_weight = 260)

}
\seealso{
The equation used here is also available in \href{https://onlinelibrary.wiley.com/doi/full/10.1111/gcb.12471}{Moraes et al. (2014)}.
}
