% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enteric_methane.R
\name{enteric_methane_lactating}
\alias{enteric_methane_lactating}
\title{Methane Emissions from lactating cows.}
\usage{
enteric_methane_lactating(dry_matter_intake_kg, ndf_perc, milk_fat_perc, bw_kg)
}
\arguments{
\item{dry_matter_intake_kg}{dry matter inatake.}

\item{ndf_perc}{Neutral detergent Fiber.}

\item{milk_fat_perc}{Milk Fat.}

\item{bw_kg}{Body Weight.}
}
\value{
g of methane per day.
}
\description{
Methane Emissions from lactating cows.
}
