% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barn_nh3.R
\name{eq_coeff}
\alias{eq_coeff}
\title{Estimates the equilibrium coefficient by the Henry's law.}
\usage{
eq_coeff(temp_c, pH)
}
\arguments{
\item{temp_c}{Temperature (°C).}

\item{pH}{pH.}
}
\value{
Equilibrium coefficient by the Henry's law
}
\description{
Estimates the equilibrium coefficient by the Henry's law.
}
\examples{
eq_coeff(temp_c = 25, pH = 7.7)

}
\seealso{
The equation used here is also available in \href{https://elibrary.asabe.org/abstract.asp?aid=21731}{Rotz and Oenenma (2005)}.
}
