% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_nitrous_oxide.R
\name{fert_manure_nitrous_oxide}
\alias{fert_manure_nitrous_oxide}
\title{Nitrous oxide emissions from fertilizer application.}
\usage{
fert_manure_nitrous_oxide(nitrogen_applied)
}
\arguments{
\item{nitrogen_applied}{Nitrogen applied (kg).}
}
\value{
Nitrous oxide emitted per kg N applied from fertilizers and manure (kg).
}
\description{
Nitrous oxide emissions from fertilizer application.
}
\examples{
fert_manure_nitrous_oxide(nitrogen_applied = 100)
}
\seealso{
The equation used here is also available in \href{https://www.ipcc-nggip.iges.or.jp/public/2006gl/pdf/4_Volume4/V4_10_Ch10_Livestock.pdf}{IPCC (2006)}.
}
