% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_aplication_nh3_emissions.R
\name{field_nh3}
\alias{field_nh3}
\title{Estimates the NH3 emissions from manure application.}
\usage{
field_nh3(tan, total_solids, days)
}
\arguments{
\item{tan}{Total ammoniacal nitrogen (kg).}

\item{total_solids}{Total solids (kg).}

\item{days}{days to incorporate manure.}
}
\value{
NH3 emissions from manure application (kg).
}
\description{
Estimates the NH3 emissions from manure application.
}
\examples{
field_nh3(tan = 7000, total_solids = 288000, days = 10)

}
\seealso{
The equation used here is also available in \href{https://www.ars.usda.gov/northeast-area/up-pa/pswmru/docs/dairy-gas-emissions-model/}{Rotz et al. (2016)}.
}
