% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manure_aplication_methane.R
\name{Fvfa_conc}
\alias{Fvfa_conc}
\title{Estimates the daily concentration of VFA in the field-applied slurry.}
\usage{
Fvfa_conc(Fvfa_ini, time_day)
}
\arguments{
\item{Fvfa_ini}{Initial VFA concentration in the slurry at application (mmol /kg slurry).}

\item{time_day}{Time since application with t = 0 representing the day of application (day).}
}
\value{
Daily concentration of VFA in the field-applied slurry ().
}
\description{
Estimates the daily concentration of VFA in the field-applied slurry.
}
\examples{
Fvfa_conc(Fvfa_ini = 0.91, time_day = 2)

}
