% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manure_aplication_methane.R
\name{Fvfa_init_conc}
\alias{Fvfa_init_conc}
\title{Estimates the initial concentration of volatile fatty acids in the slurry.}
\usage{
Fvfa_init_conc(Ftan, ph)
}
\arguments{
\item{Ftan}{concentration of TAN (total amoniacal nitrogen; NH4 + NH3) in the slurry (mmol/kg slurry).}

\item{ph}{pH of the manure slurry (dimensionless).}
}
\value{
concentration of volatile fatty acids in the slurry (mmol/kg slurry).
}
\description{
Estimates the initial concentration of volatile fatty acids in the slurry.
}
\examples{
Fvfa_init_conc(Ftan = 91.3, ph = 7)

}
