% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fossil_fuel_combustion.R
\name{gasoline_co2eq}
\alias{gasoline_co2eq}
\title{Estimates the CO2eq emissions from gasoline combustion.}
\usage{
gasoline_co2eq(gasoline_consumption)
}
\arguments{
\item{gasoline_consumption}{Gasoline consumption (L).}
}
\value{
CO2eq emissions (kg).
}
\description{
Estimates the CO2eq emissions from gasoline combustion.
}
\examples{
gasoline_co2eq(gasoline_consumption = 10)
}
