% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_inputs.R
\name{general_ui_prms}
\alias{general_ui_prms}
\title{general parameters used over the application by other modules.}
\usage{
general_ui_prms()
}
\value{
UI elements.
}
\description{
general parameters used over the application by other modules.
}
