% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enteric_methane.R
\name{gross_energy_intake}
\alias{gross_energy_intake}
\title{Estimates the gross energy intake based on dry matter intake and some diet components.}
\usage{
gross_energy_intake(
  dry_matter_intake,
  crude_protein,
  ether_extract,
  neutral_detergent_fiber
)
}
\arguments{
\item{dry_matter_intake}{Dry matter intake (kg).}

\item{crude_protein}{Diet crude protein concentration (\%).}

\item{ether_extract}{Diet ether extract concentration (\%).}

\item{neutral_detergent_fiber}{Diet neutral detergent fiber concentration (\%).}
}
\value{
Gross energy intake (MJ/day).
}
\description{
Estimates the gross energy intake based on dry matter intake and some diet components.
}
\examples{
gross_energy_intake(dry_matter_intake = 25,
                    crude_protein = 16,
                    ether_extract = 4,
                    neutral_detergent_fiber = 30)

}
\seealso{
The equation used here is also available in \href{https://onlinelibrary.wiley.com/doi/full/10.1111/gcb.12471}{Moraes et al. (2014)}.
}
