% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heifer_body_weight.R
\name{heifer_body_weight}
\alias{heifer_body_weight}
\title{Estimates the heifer mean or final body weight at given month.}
\usage{
heifer_body_weight(
  birth_weight,
  weaning_weight,
  age_first_calving,
  mature_body_weight,
  month,
  type
)
}
\arguments{
\item{birth_weight}{Calf birth weight (kg).}

\item{weaning_weight}{Weaning weight (kg).}

\item{age_first_calving}{Age at first calving (months).}

\item{mature_body_weight}{Mature body weight (kg)}

\item{month}{Month to be simulated.}

\item{type}{Weight type to be estimated: average weight in the period or final weight (kg).}
}
\value{
Heifer body weight (kg)
}
\description{
Estimates the heifer mean or final body weight at given month.
}
