% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_intake.R
\name{heifer_dry_cows_water_intake}
\alias{heifer_dry_cows_water_intake}
\title{Daily heifer and dry cows water intake.}
\usage{
heifer_dry_cows_water_intake(bw_kg)
}
\arguments{
\item{bw_kg}{Body weight (kg).}
}
\value{
Daily water intake (l).
}
\description{
Daily heifer and dry cows water intake.
}
\examples{
heifer_dry_cows_water_intake(bw_kg = 250)
}
