% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dry_matter_intake.R
\name{heifer_dry_matter_intake}
\alias{heifer_dry_matter_intake}
\title{Estimates the daily dry matter intake of heifers.}
\usage{
heifer_dry_matter_intake(mature_body_weight, body_weight)
}
\arguments{
\item{mature_body_weight}{Mature body weight (kg).}

\item{body_weight}{Body weight (kg).}
}
\value{
Heifer dry matter intake (kg).
}
\description{
Estimates the daily dry matter intake of heifers.
}
\examples{
heifer_dry_matter_intake(mature_body_weight = 680, body_weight = 300)
}
