% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potassium_excretion.R
\name{heifer_k_excretion}
\alias{heifer_k_excretion}
\title{Estimates the daily total potassium excretion of heifers.}
\usage{
heifer_k_excretion(body_weight, dry_matter_intake)
}
\arguments{
\item{body_weight}{Body weight (kg).}

\item{dry_matter_intake}{Dry matter intake (kg).}
}
\value{
Daily total potassium excretion (g).
}
\description{
Estimates the daily total potassium excretion of heifers.
}
\examples{
heifer_k_excretion(body_weight = 250, dry_matter_intake = 5.5)

}
