% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manure_excretion.R
\name{heifer_manure_excretion}
\alias{heifer_manure_excretion}
\title{Estimates the daily total manure excretion of heifers.}
\usage{
heifer_manure_excretion(dry_matter_intake, body_weight)
}
\arguments{
\item{dry_matter_intake}{Dry matter intake (kg).}

\item{body_weight}{Body weight (kg).}
}
\value{
Daily total manure excretion (kg).
}
\description{
Estimates the daily total manure excretion of heifers.
}
\examples{
heifer_manure_excretion(dry_matter_intake = 6, body_weight = 250)

}
\seealso{
The equations used here are also available in \href{https://pubmed.ncbi.nlm.nih.gov/16162547/}{Nennich et al. (2005)}.
}
