% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manure_excretion.R
\name{heifer_manure_excretion_80_to_205}
\alias{heifer_manure_excretion_80_to_205}
\title{Estimates the daily total manure excretion of heifers from 80 to 205 kg of body weight.}
\usage{
heifer_manure_excretion_80_to_205(dry_matter_intake)
}
\arguments{
\item{dry_matter_intake}{Dry matter intake (kg).}
}
\value{
Daily total manure excretion (kg).
}
\description{
Estimates the daily total manure excretion of heifers from 80 to 205 kg of body weight.
}
\examples{
heifer_manure_excretion_80_to_205(dry_matter_intake = 4)
}
