% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manure_excretion.R
\name{heifer_manure_excretion_86_to_205}
\alias{heifer_manure_excretion_86_to_205}
\title{Estimates the daily total manure excretion of heifers from 86 to 205 kg of body weight.}
\usage{
heifer_manure_excretion_86_to_205(dry_matter_intake)
}
\arguments{
\item{dry_matter_intake}{Dry matter intake (kg).}
}
\value{
Daily total manure excretion (kg).
}
\description{
Estimates the daily total manure excretion of heifers from 86 to 205 kg of body weight.
}
\examples{
heifer_manure_excretion_86_to_205(dry_matter_intake = 4)
}
\seealso{
The equation used here is also available in \href{https://pubmed.ncbi.nlm.nih.gov/16162547/}{Nennich et al. (2005)}.
}
