% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herd_projection.R
\name{herd_projection}
\alias{herd_projection}
\title{her Projection}
\usage{
herd_projection(
  time_first_calv,
  heifer_calf_born,
  stillbirth_rate,
  calves_heifers_cul,
  cow_rep_rate,
  cow_calving_int,
  n_adult_cows,
  months_to_project
)
}
\arguments{
\item{time_first_calv}{Time at first calving.}

\item{heifer_calf_born}{Heifer calf born.}

\item{stillbirth_rate}{Stillbirth rate.}

\item{calves_heifers_cul}{Culling rate for heifers and calves.}

\item{cow_rep_rate}{Culling rate of adult cows.}

\item{cow_calving_int}{Calving interval.}

\item{n_adult_cows}{Number of adult cows.}

\item{months_to_project}{Months to project.}
}
\value{
A list.
}
\description{
her Projection
}
