% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dry_matter_intake.R
\name{lactating_dry_matter_intake}
\alias{lactating_dry_matter_intake}
\title{Estimates the daily dry matter intake of lactating cows.}
\usage{
lactating_dry_matter_intake(days_milk, body_weight, milk_yield)
}
\arguments{
\item{days_milk}{days in milk.}

\item{body_weight}{Body weight (kg).}

\item{milk_yield}{Milk yield (kg).}
}
\value{
Lactating cow dry matter intake (kg).
}
\description{
Estimates the daily dry matter intake of lactating cows.
}
\examples{
lactating_dry_matter_intake(days_milk = 60, body_weight = 680, milk_yield = 40)
}
\seealso{
The equation used here is also available in \href{https://nap.nationalacademies.org/catalog/9825/nutrient-requirements-of-dairy-cattle-seventh-revised-edition-2001}{NRC (2001)}.
}
