% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enteric_methane.R
\name{lactating_enteric_methane}
\alias{lactating_enteric_methane}
\title{Estimates the daily enteric methane emission for lactating cows.}
\usage{
lactating_enteric_methane(
  gross_energy_intake,
  neutral_detergent_fiber,
  ether_extract,
  body_weight,
  milk_fat
)
}
\arguments{
\item{gross_energy_intake}{Gross energy intake (MJ/day).}

\item{neutral_detergent_fiber}{Diet neutral detergent fiber concentration (\%).}

\item{ether_extract}{Diet ether extract concentration (\%).}

\item{body_weight}{Body weight (kg).}

\item{milk_fat}{Milk fat content (\%).}
}
\value{
Lactating cow enteric methane emission (g/day).
}
\description{
Estimates the daily enteric methane emission for lactating cows.
}
\examples{
lactating_enteric_methane(gross_energy_intake = 465,
                         neutral_detergent_fiber = 30,
                         ether_extract = 4,
                         body_weight = 650,
                         milk_fat = 4)

}
\seealso{
The equation used here is also available in \href{https://onlinelibrary.wiley.com/doi/full/10.1111/gcb.12471}{Moraes et al. (2014)}.
}
