% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrogen_excretion.R
\name{lactating_n_fecal_excretion}
\alias{lactating_n_fecal_excretion}
\title{Estimates the daily fecal nitrogen excretion of lactating cows.}
\usage{
lactating_n_fecal_excretion(nitrogen_intake)
}
\arguments{
\item{nitrogen_intake}{Nitrogen intake (g/day).}
}
\value{
Daily fecal nitrogen excretion (g).
}
\description{
Estimates the daily fecal nitrogen excretion of lactating cows.
}
\examples{
lactating_n_fecal_excretion(nitrogen_intake = 640)

}
\seealso{
The equation used here is also available in \href{https://www.journalofdairyscience.org/article/S0022-0302(22)00420-9/fulltext}{Bougouin et al. (2022)}.
}
