% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphorus_excretion.R
\name{lactating_p_total_excretion}
\alias{lactating_p_total_excretion}
\title{Estimates the daily fecal phosphorous excretion of lactating cows.}
\usage{
lactating_p_total_excretion(dry_matter_intake, phosphorous_content)
}
\arguments{
\item{dry_matter_intake}{Dry matter intake (kg).}

\item{phosphorous_content}{Diet phosphorous content (\%).}
}
\value{
Daily fecal phosphorous excretion (g).
}
\description{
Estimates the daily fecal phosphorous excretion of lactating cows.
}
\examples{
lactating_p_total_excretion(dry_matter_intake = 25, phosphorous_content = 1.2)

}
