% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urine_excretion.R
\name{lactating_urine_excretion}
\alias{lactating_urine_excretion}
\title{Estimates the urine excretion for lactating cows.}
\usage{
lactating_urine_excretion(dry_matter_intake, crude_protein, milk_protein)
}
\arguments{
\item{dry_matter_intake}{Dry matter intake (kg).}

\item{crude_protein}{Diet crude protein (\%).}

\item{milk_protein}{Milk protein content(\%).}
}
\value{
Daily urine excretion for lactating cows (kg).
}
\description{
Estimates the urine excretion for lactating cows.
}
\examples{
lactating_urine_excretion(dry_matter_intake = 25, crude_protein = 16, milk_protein = 3.2)

}
