% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_intake.R
\name{lactating_water_intake}
\alias{lactating_water_intake}
\title{Daily lactating cows water intake.}
\usage{
lactating_water_intake(temp_c, milk_yield_kg, bw_kg)
}
\arguments{
\item{temp_c}{Temperature (°C).}

\item{milk_yield_kg}{Milk yield corrected for fat and protein (kg).}

\item{bw_kg}{Body weight (kg).}
}
\value{
Daily water intake (l).
}
\description{
Daily lactating cows water intake.
}
\examples{
lactating_water_intake(temp_c = 35, milk_yield_kg = 45, bw_kg = 650)
}
\seealso{
The equations used here are also available in \href{https://www.sciencedirect.com/science/article/pii/S0378377414002145}{Krauß et al. (2015)}.
}
