% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manure_excretion.R
\name{lactation_dm_manure_excretion}
\alias{lactation_dm_manure_excretion}
\title{Estimates the manure dry matter content of lactating cows.}
\usage{
lactation_dm_manure_excretion(dry_matter_intake)
}
\arguments{
\item{dry_matter_intake}{Dry matter intake (kg).}
}
\value{
Manure dry matter excretion (kg).
}
\description{
Estimates the manure dry matter content of lactating cows.
}
\examples{
lactation_dm_manure_excretion(dry_matter_intake = 25)
}
\seealso{
The equation used here is also available in \href{https://pubmed.ncbi.nlm.nih.gov/16162547/}{Nennich et al. (2005)}.
}
