% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milk_protein_content.R
\name{lambda_milk_prot}
\alias{lambda_milk_prot}
\title{Estimates the lambda hyper parameter to reach the average milk protein defined by the user.}
\usage{
lambda_milk_prot(
  obs_average_milk_prot,
  prop_primiparous,
  prop_secondiparous,
  cow_calving_interval
)
}
\arguments{
\item{obs_average_milk_prot}{Observed average milk protein (\%).}

\item{prop_primiparous}{Herd proportion of primiparous cows (\%).}

\item{prop_secondiparous}{Herd proportion of secondiparous cows (\%).}

\item{cow_calving_interval}{Calving interval (months).}
}
\value{
Lambda hyper parameter (dimensionless).
}
\description{
Estimates the lambda hyper parameter to reach the average milk protein defined by the user.
}
\examples{
lambda_milk_prot(obs_average_milk_prot = 3.5,
                 prop_primiparous = 35,
                 prop_secondiparous = 30,
                 cow_calving_interval = 14)

}
