% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milk_yield.R
\name{lambda_milk_yield}
\alias{lambda_milk_yield}
\title{Estimates the lambda hyper parameter to reach the average milk yield defined by the user.}
\usage{
lambda_milk_yield(
  obs_average_milk_yield,
  milk_freq,
  prop_primiparous,
  prop_secondiparous,
  cow_calving_interval
)
}
\arguments{
\item{obs_average_milk_yield}{Observed average milk yield (kg).}

\item{milk_freq}{Milking frequency (2 or 3).}

\item{prop_primiparous}{Herd proportion of primiparous cows (\%).}

\item{prop_secondiparous}{Herd proportion of secondiparous cows (\%).}

\item{cow_calving_interval}{Calving interval (months).}
}
\value{
Lambda hyper parameter (dimensionless).
}
\description{
Estimates the lambda hyper parameter to reach the average milk yield defined by the user.
}
\examples{
lambda_milk_yield(obs_average_milk_yield = 43,
                  milk_freq = 2,
                  prop_primiparous = 35,
                  prop_secondiparous = 30,
                  cow_calving_interval = 14)

}
