% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_carbon_dioxide.R
\name{lime_decomposition_co2}
\alias{lime_decomposition_co2}
\title{Estimates the carbon dioxide emissions from lime decomposition applied to croplands.}
\usage{
lime_decomposition_co2(lime_applied)
}
\arguments{
\item{lime_applied}{Amount of lime applied (kg).}
}
\value{
Amount of carbon dioxide emitted (kg) per kg of lime applied.
}
\description{
Estimates the carbon dioxide emissions from lime decomposition applied to croplands.
}
\examples{
lime_decomposition_co2(lime_applied = 1000)

}
\seealso{
The equation used here is also available in \href{https://www.ipcc-nggip.iges.or.jp/public/2006gl/pdf/4_Volume4/V4_11_Ch11_N2O&CO2.pdf}{IPCC, 2006}.
}
