% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manure_methane.R
\name{manure_ch4_emission_slurry}
\alias{manure_ch4_emission_slurry}
\title{Estimates the daily methane emissions from manure storage (Raw, liquid, and digestate).}
\usage{
manure_ch4_emission_slurry(vs_total, vs_d, vs_nd, temp_c, enclosed)
}
\arguments{
\item{vs_total}{Total volatile solids (kg).}

\item{vs_d}{Degradable VS in the manure (kg).}

\item{vs_nd}{Non degradable VS in the manure (kg).}

\item{temp_c}{Temperature (°C).}

\item{enclosed}{Is the manure storage enclosed: yes or no.}
}
\value{
Methane daily emissions from manure storage (kg).
}
\description{
Estimates the daily methane emissions from manure storage (Raw, liquid, and digestate).
}
\examples{
manure_ch4_emission_slurry(1, 0.417, 0.583, temp_c = 15, enclosed = "no")

}
