% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manure_density.R
\name{manure_density}
\alias{manure_density}
\title{Manure density (kg/m3).}
\usage{
manure_density(manure_dm)
}
\arguments{
\item{manure_dm}{manure dry matter content (\%)}
}
\value{
Manure density (kg/m3).
}
\description{
Manure density (kg/m3).
}
\examples{
manure_density(manure_dm = 12)
}
\seealso{
The equation used here is also available in \href{https://www.mdpi.com/2076-3417/9/13/2703}{Wang et al. (2019)}.
}
