% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_inputs.R
\name{manure_manag_ui}
\alias{manure_manag_ui}
\title{Manure managemente parameters used over the application by other modules.}
\usage{
manure_manag_ui(manure_managment)
}
\arguments{
\item{manure_managment}{Type of manure management.}
}
\value{
UI elements.
}
\description{
Manure managemente parameters used over the application by other modules.
}
