% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_inputs.R
\name{manure_ui_prms}
\alias{manure_ui_prms}
\title{Manure parameters used over the application by other modules.}
\usage{
manure_ui_prms()
}
\value{
UI elements.
}
\description{
Manure parameters used over the application by other modules.
}
