% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potassium_excretion.R
\name{milk_k_excretion}
\alias{milk_k_excretion}
\title{estimates the daily milk potassium excretion.}
\usage{
milk_k_excretion(milk_yield)
}
\arguments{
\item{milk_yield}{Milk yield (kg).}
}
\value{
Daily milk potassium excretion (g).
}
\description{
estimates the daily milk potassium excretion.
}
\examples{
milk_k_excretion(milk_yield = 45)

}
