% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_inputs.R
\name{milk_ui_prms}
\alias{milk_ui_prms}
\title{Milk UI parameters.}
\usage{
milk_ui_prms(prm)
}
\arguments{
\item{prm}{Parameter to be concatenated in the module.}
}
\value{
Milk UI components.
}
\description{
Milk UI parameters.
}
