% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrogen_excretion.R
\name{milk_urea_nitrogen}
\alias{milk_urea_nitrogen}
\title{Estimates the milk urea nitrogen (MUN).}
\usage{
milk_urea_nitrogen(nitrogen_intake)
}
\arguments{
\item{nitrogen_intake}{Nitrogen intake (g).}
}
\value{
Milk urea nitrogen (MUN) (g).
}
\description{
Estimates the milk urea nitrogen (MUN).
}
\examples{
milk_urea_nitrogen(nitrogen_intake = 640)

}
\seealso{
The equation used here is also available in \href{https://reader.elsevier.com/reader/sd/pii/S0022030215000259?token=7ECD53F63381FD64DFC3874D7B4B1BA42FF5330C206BC6F9EB59882C56F3093C1D8D1DB57FE3D563070977EDBFA8F7CC&originRegion=us-east-1&originCreation=20221102190050}{Aguirre-Villegas et al. (2015)}.
}
