% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milk_yield.R
\name{milk_yield}
\alias{milk_yield}
\title{Estimates the milk yield for lactating cows in the first, second, and third or more lactation being milked two or three times per day.}
\usage{
milk_yield(days_milk, parity, milk_freq, lambda_milk)
}
\arguments{
\item{days_milk}{Days in milk.}

\item{parity}{Parity: primiparous, secondiparous or multiparous.}

\item{milk_freq}{Milking frequency (2 or 3).}

\item{lambda_milk}{Hyper parameter to adjust the milk yield for a specific average value.}
}
\value{
Daily milk yield (kg).
}
\description{
Estimates the milk yield for lactating cows in the first, second, and third or more lactation being milked two or three times per day.
}
\examples{
milk_yield(days_milk = 35, parity = "primiparous", milk_freq = 2, lambda_milk = 1)

}
\seealso{
The equations used here are also available in \href{https://www.sciencedirect.com/science/article/pii/S0022030222004362}{Li et al. (2022)}.
}
