% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fossil_fuel_combustion.R
\name{natural_gas_co2eq}
\alias{natural_gas_co2eq}
\title{Estimates the CO2eq emissions from gas natural combustion.}
\usage{
natural_gas_co2eq(natural_gas_consumption)
}
\arguments{
\item{natural_gas_consumption}{Gas natural consumption (L).}
}
\value{
CO2eq emissions (kg).
}
\description{
Estimates the CO2eq emissions from gas natural combustion.
}
\examples{
natural_gas_co2eq(natural_gas_consumption = 10)
}
