% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphorous_loss.R
\name{phosphorous_loss}
\alias{phosphorous_loss}
\title{Estimates the phosphorous losses.}
\usage{
phosphorous_loss(p_applied, rain_fall)
}
\arguments{
\item{p_applied}{Phosphorous applied to the land (kg/year).}

\item{rain_fall}{Yearly rainfall (mm).}
}
\value{
Phosphorous loss (kg/year).
}
\description{
Estimates the phosphorous losses.
}
\examples{
phosphorous_loss(p_applied = 30, rain_fall = 4)
}
