% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barn_nh3.R
\name{resistence_nh3}
\alias{resistence_nh3}
\title{Estimates the resistance of NH3 transport from the manure surface to the free atmosphere.}
\usage{
resistence_nh3(hsc, temp_c)
}
\arguments{
\item{hsc}{Housing-specific constant (s/m).}

\item{temp_c}{Temperature (°C).}
}
\value{
Resistance of NH3 transport from the manure surface to the free atmosphere (s/m).
}
\description{
Estimates the resistance of NH3 transport from the manure surface to the free atmosphere.
}
\examples{
resistence_nh3(hsc = 260, temp_c = 25)

}
\seealso{
The equation used here is also available in \href{https://elibrary.asabe.org/abstract.asp?aid=21731}{Rotz and Oenenma (2005)}.
}
