% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starter_intake_calves.R
\name{starter_intake_calves}
\alias{starter_intake_calves}
\title{Estimates the average calf starter intake in a given day.}
\usage{
starter_intake_calves(milk_intake, age)
}
\arguments{
\item{milk_intake}{Calf milk intake (l).}

\item{age}{Calf age in days.}
}
\value{
Starter intake (kg).
}
\description{
Estimates the average calf starter intake in a given day.
}
\examples{
# Estimate the average starter intake of calves 50 days old calf consuming 6 l of milk
starter_intake_calves(milk_intake = 6, age = 50)

}
\seealso{
The equations used here are also available in \href{https://pubmed.ncbi.nlm.nih.gov/29656719/}{Silva et al. (2019)}.
}
