% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_carbon_dioxide.R
\name{urea_decomposition_co2}
\alias{urea_decomposition_co2}
\title{Estimates the carbon dioxide emissions from urea decomposition applied to croplands.}
\usage{
urea_decomposition_co2(urea_applied)
}
\arguments{
\item{urea_applied}{Amount of urea applied (kg).}
}
\value{
Amount of carbon dioxide emitted (kg) per kg of urea.
}
\description{
Estimates the carbon dioxide emissions from urea decomposition applied to croplands.
}
\examples{
urea_decomposition_co2(urea_applied = 50)

}
\seealso{
The equation used here is also available in \href{https://www.ipcc-nggip.iges.or.jp/public/2006gl/pdf/4_Volume4/V4_11_Ch11_N2O&CO2.pdf}{IPCC, 2006}.
}
