% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_cards.R
\name{value_box_spark}
\alias{value_box_spark}
\title{Customized Card.}
\usage{
value_box_spark(
  value,
  title,
  sparkobj = NULL,
  subtitle,
  info = NULL,
  icon = NULL,
  color = "aqua",
  width = 4,
  href = NULL
)
}
\arguments{
\item{value}{Value to be exposed.}

\item{title}{Card Title.}

\item{sparkobj}{Highcharter object.}

\item{subtitle}{Card subtitle.}

\item{info}{Additional information.}

\item{icon}{Icon.}

\item{color}{Color.}

\item{width}{Width.}

\item{href}{Reference.}
}
\value{
Card.
}
\description{
Customized Card.
}
