% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volatile_solids.R
\name{volatile_solids}
\alias{volatile_solids}
\title{Estimates the daily total volatile solids.}
\usage{
volatile_solids(dry_matter_intake, crude_protein, neutral_detergent_fiber)
}
\arguments{
\item{dry_matter_intake}{Dry matter intake (kg).}

\item{crude_protein}{Diet crude protein (\%).}

\item{neutral_detergent_fiber}{Diet neutral detergent fiber (\%).}
}
\value{
Daily total volatile solids (kg).
}
\description{
Estimates the daily total volatile solids.
}
\examples{
volatile_solids(dry_matter_intake = 25,
               crude_protein = 16,
               neutral_detergent_fiber = 30)

}
\seealso{
The equation used here is also available in \href{https://nap.nationalacademies.org/catalog/25806/nutrient-requirements-of-dairy-cattle-eighth-revised-edition}{NASEM (2021)}.
}
