% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wisconsin_weather_data}
\alias{wisconsin_weather_data}
\title{Winsconsin county weather dataset.}
\format{
Dataset with 26280 rows and 6 columns:
\describe{
  \item{county}{County name.}
  \item{yday}{Year day.}
  \item{precip_mm}{Precipitation (mm).}
  \item{max_tempC}{Maximum temperature (°C).}
  \item{min_tempC}{Minimum temperature (°C).}
  \item{aver_tempC}{Average temperature (°C).}
  ...
}
}
\source{
Data retrieved from \href{https://daymet.ornl.gov/}{Daymet website}.
}
\usage{
wisconsin_weather_data
}
\description{
This dataset contains weather information summarized for each county in Wisconsin since 1990 to 2021.
}
\keyword{datasets}
