% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milk_protein_content.R
\name{average_milk_protein}
\alias{average_milk_protein}
\title{Estimates the average milk protein between two days in milk.}
\usage{
average_milk_protein(parity, day_min, day_max)
}
\arguments{
\item{parity}{Parity: primiparous, secondiparous, or multiparous.}

\item{day_min}{Initial day in milk.}

\item{day_max}{Final day in milk.}
}
\value{
Average milk protein between two days in milk (\%).
}
\description{
Estimates the average milk protein between two days in milk.
}
\examples{
average_milk_protein(parity = "multiparous", day_min = 30, day_max = 60)

}
