% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrogen_excretion.R
\name{dry_fecal_n_excretion}
\alias{dry_fecal_n_excretion}
\title{Estimates the daily fecal nitrogen excretion of dry cows.}
\usage{
dry_fecal_n_excretion(nitrogen_intake)
}
\arguments{
\item{nitrogen_intake}{Nitrogen intake (g/day).}
}
\value{
Daily fecal nitrogen excretion (g).
}
\description{
Estimates the daily fecal nitrogen excretion of dry cows.
}
\examples{
dry_fecal_n_excretion(nitrogen_intake = 230)

}
