% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphorus_excretion.R
\name{lactating_p_total_excretion}
\alias{lactating_p_total_excretion}
\title{Estimates the daily fecal phosphorous excretion of lactating cows.}
\usage{
lactating_p_total_excretion(
  dry_matter_intake,
  phosphorous_content,
  milk_p_excretion
)
}
\arguments{
\item{dry_matter_intake}{Dry matter intake (kg).}

\item{phosphorous_content}{Diet phosphorous content (\%).}

\item{milk_p_excretion}{Milk phosphorous excretion (g).}
}
\value{
Daily fecal phosphorous excretion (g).
}
\description{
Estimates the daily fecal phosphorous excretion of lactating cows.
}
\note{
This function assumes no tissue mobilization or retention of phosphorous by the cows.
}
\examples{
lactating_p_total_excretion(dry_matter_intake = 25,
                            phosphorous_content = 0.4,
                            milk_p_excretion = 45)

}
\seealso{
The equation used here is also available in \href{https://pubmed.ncbi.nlm.nih.gov/16162547/}{Nennich et al. (2005)}.
}
