% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manure_methane.R
\name{manure_ch4_emission_slurry}
\alias{manure_ch4_emission_slurry}
\title{Estimates the daily methane emissions from manure storage (Raw, liquid, and digestate).}
\usage{
manure_ch4_emission_slurry(
  volatile_solids_total,
  volatile_solids_d,
  volatile_solids_nd,
  temp_c,
  enclosed
)
}
\arguments{
\item{volatile_solids_total}{Total volatile solids (kg).}

\item{volatile_solids_d}{Degradable VS in the manure (kg).}

\item{volatile_solids_nd}{Non degradable VS in the manure (kg).}

\item{temp_c}{Temperature (°C).}

\item{enclosed}{Is the manure storage enclosed: yes or no.}
}
\value{
Methane daily emissions from manure storage (kg).
}
\description{
Estimates the daily methane emissions from manure storage (Raw, liquid, and digestate).
}
\examples{
manure_ch4_emission_slurry(volatile_solids_total = 1,
                           volatile_solids_d = 0.417,
                           volatile_solids_nd = 0.583,
                           temp_c = 15,
                           enclosed = "no")

}
\seealso{
The equation used here is also available in \href{https://elibrary.asabe.org/abstract.asp?aid=27781}{Chianese et al. (2009)}.
}
